[feat] 마커 애니메이션을 위한 딕셔너리 추가 및 리펙토링
