+++ NaverMapA/NaverMapA/Controller/MainViewController+NMFMapViewCameraDelegate.swift
@@ -1,0 +1,25 @@
+//
+//  ViewController+NMFMapViewCameraDelegate.swift
+//  NaverMapA
+//
+//  Created by 박태희 on 2020/12/01.
+//
+
+import Foundation
+import NMapsMap
+
+extension MainViewController: NMFMapViewCameraDelegate {
+    
+    func mapViewCameraIdle(_ mapView: NMFMapView) {
+        let coordBounds = mapView.projection.latlngBounds(fromViewBounds: UIScreen.main.bounds)
+        DispatchQueue.global().async {
+            let bounds = CoordinateBounds(southWestLng: coordBounds.southWestLng,
+                                          northEastLng: coordBounds.northEastLng,
+                                          southWestLat: coordBounds.southWestLat,
+                                          northEastLat: coordBounds.northEastLat)
+            
+            let places = self.dataProvider.fetch(bounds: bounds)
+            self.viewModel?.updatePlaces(places: places, bounds: bounds)
+        }
+    }
+}
