+++ NaverMapA/NaverMapA/Controller/MainViewController.swift
@@ -16,7 +16,7 @@ class MainViewController: UIViewController {
     var clusterMarkers = [NMFMarker]()
     var beforeClusterMarkers = [NMFMarker]()
     var clusterMarkersCount = 0
-    var zoomLevel: Double = 18 {
+    var prevZoomLevel: Double = 18 {
         didSet(oldValue) {
             if oldValue != mapView.zoomLevel && clusterMarkersCount != clusterMarkers.count {
                 clusterMarkersCount = clusterMarkers.count
@@ -24,7 +24,7 @@ class MainViewController: UIViewController {
             }
         }
     }
-    private lazy var dataProvider: PlaceProvider = {
+    lazy var dataProvider: PlaceProvider = {
         let provider = PlaceProvider.shared
         provider.fetchedResultsController.delegate = self
         return provider
@@ -73,7 +73,7 @@ class MainViewController: UIViewController {
                         marker.mapView = self.mapView
                         self.clusterMarkers.append(marker)
                     }
-                    self.zoomLevel = self.mapView.zoomLevel
+                    self.prevZoomLevel = self.mapView.zoomLevel
                 }
             })
         }
@@ -140,21 +140,5 @@ class MainViewController: UIViewController {
 
 extension MainViewController: NSFetchedResultsControllerDelegate {
     func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-    }
-}
-
-extension MainViewController: NMFMapViewCameraDelegate {
-    
-    func mapViewCameraIdle(_ mapView: NMFMapView) {
-        let coordBounds = mapView.projection.latlngBounds(fromViewBounds: UIScreen.main.bounds)
-        DispatchQueue.global().async {
-            let bounds = CoordinateBounds(southWestLng: coordBounds.southWestLng,
-                                          northEastLng: coordBounds.northEastLng,
-                                          southWestLat: coordBounds.southWestLat,
-                                          northEastLat: coordBounds.northEastLat)
-            
-            let places = self.dataProvider.fetch(bounds: bounds)
-            self.viewModel?.updatePlaces(places: places, bounds: bounds)
-        }
     }
 }
